Fix calculation of Y axis when stacked is true
Just taking the max of all sources and adding it up leads to too large a
range, so we need to go through all the items from all sources, and
determine which of the total items is the max. Unfortunately this leads
to yet-another-loop-over-items, but I currently do not have a better
solution.