Trick doxygen into placing all classes into the right namespace
doxyqml places everything in a namespace corresponding to the module
path and unfortunately has no way of disabling that behaviour. So
instead, trick doxygen to place all the c++ items also in a namespace
matching the module path.

This is done only for doxygen because the module itself does not really
need namespacing and Qt/moc can get nasty when dealing with namespaces.