Line charts: Reduce the number of points per segment
Apparently, the Intel graphics driver will use the CPU for shaders that
it deems "too complex". The 100 point count was hitting this number, so
reduce it so things are done by the cpu again.

Maybe at some point in the future this value should be based on the
current GPU.