Rework PieChart implementation
This makes use of the new sdf_torus_segment sdf function to render pie
chart segments in one go. This simplifies the shader greatly, removing
almost all need for preprocessing the pie segments in the pie chart node
and removing a double loop in the shader. The tradeoff is that the code
for rendering the actual segment is a little more complex.