Fix controls plugin loading
Prior to fc43200f80533534e3eaf62b9630aa5beb03c83f, the plugin was called
"chartscontrolsplugin", but that change renamed it.

Restore the name to ensure the plugin is actually found.

Tested with `plasmoidviewer -a org.kde.plasma.networkmanagement`. It would not display properly before (because it would not find the quickcharts controls module), now it does.