LineChart: Replace shader clipping with geometry-based clipping
Apparently shader-based clipping can cause artifacts on some GPUs
(Intel). So instead of rendering the full height rect for a line chart
segment and discarding fragments in the shader, resize the geometry of
the SDF part so it is only used for the relevant part, anything above is
discarded and anything below is rendered using a simple colored
rectangle.