controls: Ceil all components of preferredWidth in LegendDelegate contents
Layout internally `ceil()`s all sizes, so not doing the same means we
get a slight mismatch between the preferred size the layout wants to be
at and the preferred size we calculate, leading to the label being
elided.

While we're at it, drop the Item container and FontMetrics usage which
can be done using Label's implicitWidth these days.

BUG: 486411